reload video index if vfs directory has changed Tidy VP ( VP5 needs work)

+++ src/rpi/screen.h
@@ -12,7 +12,7 @@ void screen_create_RGB_plane( uint32_t planeno, uint32_t width , uint32_t height
 void screen_release_plane( uint32_t planeno );
 void screen_set_plane_position( uint32_t planeno, int x, int y );
 void screen_plane_enable(uint32_t planeno, bool enable);
-void screen_plane_setalpha( uint32_t planeno , uint8_t alpha );
+void screen_plane_setalpha( uint32_t planeno , int alpha );
 void screen_update_palette_entry(uint32_t entry, uint32_t r, uint32_t g, uint32_t b);
 void screen_set_palette(uint32_t planeno, uint32_t palette, uint32_t flags);
 uint32_t screen_get_palette_entry( uint32_t entry );
